Initial conversion of packaging script.
This is still a rough script to package up BTM in preparation for its
sharing on ArcGIS Online, and as a GitHub release. It still doesn't do
many of the things we'd like it to do, such as handling Git tagging, but
it's a starting point and produces consistent builds.